Pulled db script into a separate file. New test blob.mp4
